+++ Jewelry Shop_Front-End/src/pages/SearchPage.jsx
@@ -1,42 +1,52 @@
 import styled from "styled-components";
 import Navbar from "../components/Navbar";
-import Categories from "../components/Categories";
-import Slider from "../components/Slider";
-import ElementOne from "../components/ElementOne";
-import ListProduct from "../components/ListProduct";
-import Album from "../components/Album";
 import Footer from "../components/Footer";
 import RiseLoader from "react-spinners/RiseLoader";
-import Aos from "aos";
 import "aos/dist/aos.css";
-import SearchpageBody from '../components/searchpage/SearchpageBody'
+import SearchpageBody from "../components/searchpage/SearchpageBody";
 import { useEffect, useState } from "react";
+import axios from "axios";
+import { useParams } from "react-router-dom";
 
 // const Container = styled.div``;
-const Spinner = styled.div`
-  height: 100%;
-  flex-direction: column;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
+// const Spinner = styled.div`
+//   height: 100%;
+//   flex-direction: column;
+//   display: flex;
+//   align-items: center;
+//   justify-content: center;
+// `;
 const Container = styled.div`
-   font-family: 'Jost', sans-serif;
+  font-family: "Jost", sans-serif;
 `;
 
 function SearchPage() {
-  
   const [loading, setLoading] = useState(false);
+  const { searchtext } = useParams();
+  const [foundProducts, setFoundProducts] = useState([]);
+
   useEffect(() => {
     setLoading(true);
-    setTimeout(() => {
-      setLoading(false);
-    }, 500);
-  }, []);
+    async function fetchData() {
+      try {
+        const response = await axios.get(
+          `http://localhost:9999/api/v1/products/search/${searchtext}`
+        );
+        const data = response.data.data;
+        setFoundProducts(data);
+        setLoading(false);
+      } catch (error) {
+        setLoading(false);
+        console.error("Error fetching data:", error);
+      }
+    }
+
+    fetchData();
+  }, [searchtext]);
   return (
     <>
       {loading ? (
-        <Container 
+        <Container
           style={{
             height: "100vh",
             display: "flex",
@@ -50,11 +60,11 @@ function SearchPage() {
       ) : (
         <Container>
           <Navbar />
-          
-          <SearchpageBody />
-        
+          <SearchpageBody
+            searchtext={searchtext}
+            foundProducts={foundProducts}
+          />
           <Footer />
-          
         </Container>
       )}
     </>
